third_party/WebKit/LayoutTests/fast/events/pointerevents/mouse-pointer-event-properties.html

<!DOCTYPE HTML>
<script src="../../../resources/js-test.js"></script>
<script src="../resources/input-modifiers.js"></script>
<style>
div.box {
  margin: 10px;
  padding: 50px;
  float: left;
}
</style>

<div id="target" class="box" style="background-color:red">
</div>

<div id="console"></div>

<script>
window.name = "mainWindow";
description("Verifies that pointer event parameters are correct when fired for mouse events.");

var testEventList = ["mouseenter", "mouseleave", "mouseover", "mouseout", "mouseup", "mousedown", "mousemove",
                   "pointerenter", "pointerleave", "pointerover", "pointerout", "pointerup", "pointerdown", "pointermove"];
var lastPointerEvents = [];

var checkKeyModifiers = false;

var numericAttributes = [
    "clientX",
    "clientY",
    "layerX",
    "layerY",
    "movementX",
    "movementY",
    "offsetX",
    "offsetY",
    "pageX",
    "pageY",
    "screenX",
    "screenY",
    "x",
    "y",
    "button",
    "buttons",
    "pressure",
    "width",
    "height",
];

function init() {
  var targetDiv = document.getElementById("target");

  testEventList.forEach(function(eventName) {

    targetDiv.addEventListener(eventName, function(event) {

      debug("Received " + event.type);

      if (event.type.startsWith("pointer"))
        lastPointerEvents.push(event);
      else {
        if (event.type == "mouseout" || event.type == "mouseover") {
          shouldBeEqualToNumber("lastPointerEvents.length", 2);
        } else {
          shouldBeEqualToNumber("lastPointerEvents.length", 1);
        }
        shouldBeEqualToString("lastPointerEvents[0].type", event.type.replace("mouse", "pointer"));

        if (!checkKeyModifiers) {

          if (lastPointerEvents[0].type=="pointerenter" || lastPointerEvents[0].type=="pointerleave") {
            shouldBeFalse("lastPointerEvents[0].bubbles");
            shouldBeFalse("lastPointerEvents[0].cancelable");
          } else {
            shouldBeTrue("lastPointerEvents[0].bubbles");
            shouldBeTrue("lastPointerEvents[0].cancelable");
          }

          shouldBeEqualToNumber("lastPointerEvents[0].pointerId", 1);
          shouldBeEqualToString("lastPointerEvents[0].pointerType", "mouse");
          shouldBeTrue("lastPointerEvents[0].isPrimary");

          numericAttributes.forEach(function(attribute) {
            var expectedValue = eval("event." + attribute);
            if (attribute == "button" && event.type != "mousedown" && event.type != "mouseup")
              expectedValue = -1;
            if (attribute == "width" || attribute == "height")
              expectedValue = 1;
            if (attribute == "pressure") {
              if (event.buttons == 0)
                expectedValue = 0.0;
              else
                expectedValue = 0.5;
            }
            shouldBeEqualToNumber("lastPointerEvents[0]." + attribute, expectedValue);
          });
          shouldBeEqualToString("lastPointerEvents[0].view.name", "mainWindow");

        } else {

          forEachModifier(function(attr, modifierName) {
            var getModifierStateStr = ".getModifierState('" + modifierName + "');"
            if (eval("event" + getModifierStateStr))
              shouldBeTrue("lastPointerEvents[0]" + getModifierStateStr);
            else
              shouldBeFalse("lastPointerEvents[0]" + getModifierStateStr);
          });

        }

        lastPointerEvents.splice(0, 1);
      }
    });

  });
}

function runTests() {
  var rect = document.getElementById("target").getBoundingClientRect();

  debug("--- move mouse into target ---");
  eventSender.mouseMoveTo(rect.left + 5, rect.top + 5);
  debug("");

  debug("--- move within target ---");
  eventSender.mouseMoveTo(rect.left + 7, rect.top + 15);
  eventSender.mouseMoveTo(rect.left + 5, rect.top + 5);
  debug("");

  debug("--- click each button ---");
  for (var button = 0; button <=2; button++) {
    eventSender.mouseDown(button);
    eventSender.mouseUp(button);
    // TODO(crbug.com/548226): Investigate missing events in win_chromium_rel_ng on 3rd down/up.
  }
  debug("");

  debug("--- click with each modifier ---");
  checkKeyModifiers = true;
  forEachModifier(function(attr, modifierName, eventSenderName) {
    eventSender.mouseDown(0, [eventSenderName]);
    eventSender.mouseUp(0, [eventSenderName]);
  });
  checkKeyModifiers = false;
  debug("");

  debug("--- move mouse out of target ---");
  eventSender.mouseMoveTo(rect.left - 5, rect.top - 5);
}

init();
if (window.eventSender)
  runTests();
else
  debug("This test requires eventSender");

</script>
